Move rebalance liquidity/target methods into ExchangeContext trait
